requirements.txt
mysql.connector
playsound
getpass
hashlib

